Refactor onboarding wizard into step components
Split the 670-line OnboardingWizard into a ~290-line container plus per-step
presentational components, with no behavior change:

- Extract each STEP_ORDER step into src/components/Onboarding/steps/*Step.tsx,
  each taking the shared `w` (useOnboardingDraft) bag; the wizard body is now
  eight one-line conditionals
- Add a reusable CopyableUrlRow (label + read-only URL + Copy + help text),
  replacing the two duplicated post-publish blocks and removing the IIFE
- Move ReviewSummary into its own file
- Export OnboardingDraft = ReturnType<typeof useOnboardingDraft> for step props;
  drop the unused `total` from the hook return (keep `index`, used by canGoBack)
- Move the highest-churn inline styles to .onboarding-*/.review-* CSS classes

The container keeps the step-gated effects, publish handler, and chrome.
